refactor(pipeline): combine action provider and extMarketSvc
