+++ modules/pipeline/pkg/action_info/passedDataWhenCreate.go
@@ -26,8 +26,8 @@ import (
 
 	"github.com/erda-project/erda/apistructs"
 	"github.com/erda-project/erda/bundle"
+	"github.com/erda-project/erda/modules/pipeline/providers/actionmgr"
 	"github.com/erda-project/erda/modules/pipeline/services/apierrors"
-	"github.com/erda-project/erda/modules/pipeline/services/extmarketsvc"
 	"github.com/erda-project/erda/modules/pipeline/spec"
 	"github.com/erda-project/erda/pkg/parser/diceyml"
 	"github.com/erda-project/erda/pkg/parser/pipelineyml"
@@ -39,7 +39,7 @@ type PassedDataWhenCreate struct {
 	bdl              *bundle.Bundle
 	actionJobDefines *sync.Map
 	actionJobSpecs   *sync.Map
-	extMarketSvc     *extmarketsvc.ExtMarketSvc
+	actionMgr        actionmgr.Interface
 }
 
 func (that *PassedDataWhenCreate) GetActionJobDefine(actionTypeVersion string) *diceyml.Job {
@@ -75,7 +75,7 @@ func (that *PassedDataWhenCreate) GetActionJobSpecs(actionTypeVersion string) *a
 	return nil
 }
 
-func (that *PassedDataWhenCreate) InitData(bdl *bundle.Bundle, extMarketSvc *extmarketsvc.ExtMarketSvc) {
+func (that *PassedDataWhenCreate) InitData(bdl *bundle.Bundle, actionMgr actionmgr.Interface) {
 	if that == nil {
 		return
 	}
@@ -86,7 +86,7 @@ func (that *PassedDataWhenCreate) InitData(bdl *bundle.Bundle, extMarketSvc *ext
 	if that.actionJobSpecs == nil {
 		that.actionJobSpecs = &sync.Map{}
 	}
-	that.extMarketSvc = extMarketSvc
+	that.actionMgr = actionMgr
 	that.bdl = bdl
 }
 
@@ -102,18 +102,18 @@ func (that *PassedDataWhenCreate) PutPassedDataByPipelineYml(pipelineYml *pipeli
 				if action.Type.IsSnippet() {
 					continue
 				}
-				extItem := extmarketsvc.MakeActionTypeVersion(action)
+				extItem := that.actionMgr.MakeActionTypeVersion(action)
 				// extension already searched, skip
 				if _, ok := that.actionJobDefines.Load(extItem); ok {
 					continue
 				}
-				extItems = append(extItems, extmarketsvc.MakeActionTypeVersion(action))
+				extItems = append(extItems, that.actionMgr.MakeActionTypeVersion(action))
 			}
 		}
 	}
 
 	extItems = strutil.DedupSlice(extItems, true)
-	actionJobDefines, actionJobSpecs, err := that.extMarketSvc.SearchActions(extItems, extmarketsvc.MakeActionLocationsBySource(p.PipelineSource))
+	actionJobDefines, actionJobSpecs, err := that.actionMgr.SearchActions(extItems, that.actionMgr.MakeActionLocationsBySource(p.PipelineSource))
 	if err != nil {
 		return apierrors.ErrCreatePipelineGraph.InternalError(err)
 	}
