+++ modules/pipeline/providers/actionmgr/provider.go
@@ -1,0 +1,82 @@
+// Copyright (c) 2021 Terminus, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package actionmgr
+
+import (
+	"context"
+	"reflect"
+	"sync"
+	"time"
+
+	"github.com/erda-project/erda-infra/base/logs"
+	"github.com/erda-project/erda-infra/base/servicehub"
+	"github.com/erda-project/erda-infra/pkg/transport"
+	"github.com/erda-project/erda-infra/providers/mysqlxorm"
+	"github.com/erda-project/erda-proto-go/core/pipeline/action/pb"
+	"github.com/erda-project/erda/apistructs"
+	"github.com/erda-project/erda/bundle"
+	"github.com/erda-project/erda/modules/pipeline/providers/actionmgr/db"
+	"github.com/erda-project/erda/pkg/common/apis"
+	"github.com/erda-project/erda/pkg/goroutinepool"
+)
+
+type config struct {
+	RefreshInterval time.Duration `file:"refresh_interval" default:"1m"`
+	PoolSize        int           `file:"pool_size" default:"20"`
+}
+
+// +provider
+type provider struct {
+	Cfg      *config
+	Log      logs.Logger
+	Register transport.Register
+	MySQL    mysqlxorm.Interface `autowired:"mysql-xorm"`
+
+	sync.Mutex
+	bdl *bundle.Bundle
+	*actionService
+
+	actions        map[string]apistructs.ExtensionVersion
+	defaultActions map[string]apistructs.ExtensionVersion
+	pools          *goroutinepool.GoroutinePool
+}
+
+func (s *provider) Init(ctx servicehub.Context) error {
+	s.actionService = &actionService{s, &db.Client{Interface: s.MySQL}}
+	if s.Register != nil {
+		pb.RegisterActionServiceImp(s.Register, s.actionService, apis.Options())
+	}
+	s.actions = make(map[string]apistructs.ExtensionVersion)
+	s.defaultActions = make(map[string]apistructs.ExtensionVersion)
+	s.pools = goroutinepool.New(s.Cfg.PoolSize)
+	s.bdl = bundle.New(bundle.WithAllAvailableClients())
+	return nil
+}
+
+func (s *provider) Run(ctx context.Context) error {
+	s.continuousRefreshAction()
+	return nil
+}
+
+func init() {
+	interfaceType := reflect.TypeOf((*Interface)(nil)).Elem()
+	servicehub.Register("actionmgr", &servicehub.Spec{
+		Types:                []reflect.Type{interfaceType},
+		OptionalDependencies: []string{"service-register"},
+		Description:          "pipeline action mgr",
+		ConfigFunc:           func() interface{} { return &config{} },
+		Creator:              func() servicehub.Provider { return &provider{} },
+	})
+}
