+++ modules/pipeline/providers/cache/provider.go
@@ -23,22 +23,22 @@ import (
 	"github.com/erda-project/erda-infra/providers/mysqlxorm"
 	"github.com/erda-project/erda/bundle"
 	"github.com/erda-project/erda/modules/pipeline/dbclient"
-	"github.com/erda-project/erda/modules/pipeline/services/extmarketsvc"
+	"github.com/erda-project/erda/modules/pipeline/providers/actionmgr"
 )
 
 type config struct {
 }
 
 type provider struct {
-	Log   logs.Logger
-	Cfg   *config
-	MySQL mysqlxorm.Interface
+	Log       logs.Logger
+	Cfg       *config
+	MySQL     mysqlxorm.Interface
+	ActionMgr actionmgr.Interface
 
 	dbClient *dbclient.Client
 	bdl      *bundle.Bundle
 
-	cacheMap     sync.Map
-	extMarketSvc *extmarketsvc.ExtMarketSvc
+	cacheMap sync.Map
 }
 
 func (p *provider) Init(ctx servicehub.Context) error {
