+++ modules/pipeline/providers/reconciler/interface.go
@@ -20,7 +20,6 @@ import (
 	"github.com/erda-project/erda/apistructs"
 	"github.com/erda-project/erda/modules/pipeline/pkg/action_info"
 	"github.com/erda-project/erda/modules/pipeline/services/actionagentsvc"
-	"github.com/erda-project/erda/modules/pipeline/services/extmarketsvc"
 	"github.com/erda-project/erda/modules/pipeline/spec"
 	"github.com/erda-project/erda/pkg/parser/pipelineyml"
 )
@@ -29,7 +28,7 @@ type Interface interface {
 	ReconcileOnePipeline(ctx context.Context, pipelineID uint64)
 
 	// InjectLegacyFields TODO decouple it
-	InjectLegacyFields(f *PipelineSvcFuncs, actionAgentSvc *actionagentsvc.ActionAgentSvc, extMarketSvc *extmarketsvc.ExtMarketSvc)
+	InjectLegacyFields(f *PipelineSvcFuncs, actionAgentSvc *actionagentsvc.ActionAgentSvc)
 }
 
 type PipelineSvcFuncs struct {
@@ -41,8 +40,7 @@ type PipelineSvcFuncs struct {
 	PreCheck                                func(p *spec.Pipeline, stages []spec.PipelineStage, userID string, autoRun bool) error
 }
 
-func (r *provider) InjectLegacyFields(f *PipelineSvcFuncs, actionAgentSvc *actionagentsvc.ActionAgentSvc, extMarketSvc *extmarketsvc.ExtMarketSvc) {
+func (r *provider) InjectLegacyFields(f *PipelineSvcFuncs, actionAgentSvc *actionagentsvc.ActionAgentSvc) {
 	r.pipelineSvcFuncs = f
 	r.actionAgentSvc = actionAgentSvc
-	r.extMarketSvc = extMarketSvc
 }
