+++ modules/pipeline/providers/reconciler/provider.go
@@ -24,6 +24,7 @@ import (
 	"github.com/erda-project/erda-infra/providers/mysqlxorm"
 	"github.com/erda-project/erda/bundle"
 	"github.com/erda-project/erda/modules/pipeline/dbclient"
+	"github.com/erda-project/erda/modules/pipeline/providers/actionmgr"
 	"github.com/erda-project/erda/modules/pipeline/providers/cache"
 	"github.com/erda-project/erda/modules/pipeline/providers/clusterinfo"
 	"github.com/erda-project/erda/modules/pipeline/providers/cron/compensator"
@@ -32,7 +33,6 @@ import (
 	"github.com/erda-project/erda/modules/pipeline/providers/reconciler/taskpolicy"
 	"github.com/erda-project/erda/modules/pipeline/providers/resourcegc"
 	"github.com/erda-project/erda/modules/pipeline/services/actionagentsvc"
-	"github.com/erda-project/erda/modules/pipeline/services/extmarketsvc"
 )
 
 type provider struct {
@@ -47,6 +47,7 @@ type provider struct {
 	EdgeRegister    edgepipeline_register.Interface
 	ResourceGC      resourcegc.Interface
 	CronCompensator compensator.Interface
+	ActionMgr       actionmgr.Interface
 
 	dbClient *dbclient.Client
 	bdl      *bundle.Bundle
@@ -54,7 +55,6 @@ type provider struct {
 	// legacy fields
 	pipelineSvcFuncs *PipelineSvcFuncs
 	actionAgentSvc   *actionagentsvc.ActionAgentSvc
-	extMarketSvc     *extmarketsvc.ExtMarketSvc
 }
 
 type config struct {
