+++ modules/pipeline/providers/reconciler/taskrun/taskrun.go
@@ -24,10 +24,10 @@ import (
 	"github.com/erda-project/erda/modules/pipeline/commonutil/costtimeutil"
 	"github.com/erda-project/erda/modules/pipeline/dbclient"
 	"github.com/erda-project/erda/modules/pipeline/pipengine/actionexecutor/types"
+	"github.com/erda-project/erda/modules/pipeline/providers/actionmgr"
 	"github.com/erda-project/erda/modules/pipeline/providers/clusterinfo"
 	"github.com/erda-project/erda/modules/pipeline/providers/edgepipeline_register"
 	"github.com/erda-project/erda/modules/pipeline/services/actionagentsvc"
-	"github.com/erda-project/erda/modules/pipeline/services/extmarketsvc"
 	"github.com/erda-project/erda/modules/pipeline/spec"
 )
 
@@ -57,7 +57,7 @@ type TaskRun struct {
 
 	// svc
 	ActionAgentSvc *actionagentsvc.ActionAgentSvc
-	ExtMarketSvc   *extmarketsvc.ExtMarketSvc
+	ActionMgr      actionmgr.Interface
 
 	RetryInterval time.Duration
 }
@@ -67,7 +67,7 @@ type TaskRun struct {
 func New(ctx context.Context, task *spec.PipelineTask,
 	executor types.ActionExecutor, p *spec.Pipeline, bdl *bundle.Bundle, dbClient *dbclient.Client,
 	actionAgentSvc *actionagentsvc.ActionAgentSvc,
-	extMarketSvc *extmarketsvc.ExtMarketSvc, clusterInfo clusterinfo.Interface, edgeRegister edgepipeline_register.Interface,
+	actionMgr actionmgr.Interface, clusterInfo clusterinfo.Interface, edgeRegister edgepipeline_register.Interface,
 	retryInterval time.Duration,
 ) *TaskRun {
 	// make executor has buffer, don't block task framework
@@ -92,7 +92,7 @@ func New(ctx context.Context, task *spec.PipelineTask,
 		ExecutorDoneCh: executorCh,
 
 		ActionAgentSvc: actionAgentSvc,
-		ExtMarketSvc:   extMarketSvc,
+		ActionMgr:      actionMgr,
 
 		RetryInterval: retryInterval,
 	}
