+++ modules/pipeline/services/pipelinesvc/create_test.go
@@ -25,6 +25,7 @@ import (
 	"github.com/erda-project/erda/apistructs"
 	"github.com/erda-project/erda/modules/pipeline/dbclient"
 	"github.com/erda-project/erda/modules/pipeline/pkg/action_info"
+	"github.com/erda-project/erda/modules/pipeline/providers/actionmgr"
 	"github.com/erda-project/erda/modules/pipeline/spec"
 	"github.com/erda-project/erda/pkg/parser/pipelineyml"
 )
@@ -346,7 +347,7 @@ func TestPipelineSvc_getYmlActionTasks(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			s := &PipelineSvc{}
+			s := &PipelineSvc{actionMgr: &actionmgr.MockActionMgr{}}
 
 			var pipelineYml *pipelineyml.PipelineYml
 			var err error
@@ -504,7 +505,7 @@ func TestPipelineSvc_MergePipelineYmlTasks(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			s := &PipelineSvc{}
+			s := &PipelineSvc{actionMgr: &actionmgr.MockActionMgr{}}
 			yml, _ := pipelineyml.New([]byte(tt.args.pipelineYml))
 			_, err := s.MergePipelineYmlTasks(yml, tt.args.dbTasks, tt.args.p, tt.args.dbStages, tt.args.passedDataWhenCreate)
 			if (err != nil) != tt.wantErr {
